+++ ost_visualizer/presentation/dialogs/job_statuses_dialog.py
@@ -10,6 +10,7 @@ from ..config import (
 )
 from ..dtos.picker_dialog_result_dto import PickerDialogResult
 from ..utils.dialog import BasePickerDialog, ItemRecord
+from ..utils.tree_widget import set_tree_item_row_height
 
 
 class _StatusRecord(ItemRecord):
@@ -72,8 +73,7 @@ class JobStatusesDialog(BasePickerDialog):
 
     def _add_tree_item(self, record: _StatusRecord) -> QtWidgets.QTreeWidgetItem:
         item = QtWidgets.QTreeWidgetItem(["", record["name"]])
-        item.setSizeHint(0, QtCore.QSize(0, 26))
-        item.setSizeHint(1, QtCore.QSize(0, 26))
+        set_tree_item_row_height(item, self.tree.columnCount())
         item.setData(1, self._UID_ROLE, record["uid"])
         item.setFlags(item.flags() | QtCore.Qt.ItemFlag.ItemIsEditable)
         self.tree.addTopLevelItem(item)